backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
django-filter==24.3
gunicorn==23.0.0
faker
python-dotenv==1.2.2
djangorestframework-simplejwt
